Separated draft controller
